requirements.txt
streamlit==1.31.0
pandas==2.2.0
plotly==5.18.0
openpyxl==3.1.2
reportlab==4.0.8
python-dateutil==2.8.2
xlsxwriter==3.1.9
Pillow==10.2.0
